A bunch more work towards making repo connector what it should be for everything to work together nicely.
